Extract HasHome helper
