+++ src/Compilers/CSharp/Test/Emit/CodeGen/CodeGenInParametersTests.cs
@@ -336,8 +336,8 @@ GetField 1
   IL_0021:  ret
 }");
         }
-        [Fact]
 
+        [Fact]
         public void InIndexerReorderWithCopy()
         {
             var verifier = CompileAndVerify(@"
